refactor ICRUD_Operations into  ICrudOperations\IQueryOperations
